Updated JS
Fixed Decimal Problem

+++ InequalityCaluclatorWebApp/script.js
@@ -4,12 +4,12 @@ var calcButton = document.getElementById("calcButton");
 // Declare calcEquation function
 function calcEquation() {
 	// Connect parameters
-	var val1 = parseInt(document.getElementById("val1").value);
-	var val2 = parseInt(document.getElementById("val2").value);
-	var inequalityVal = parseInt(document.getElementById("inequalityVal").value);
+	var val1 = parseFloat(document.getElementById("val1").value);
+	var val2 = parseFloat(document.getElementById("val2").value);
+	var inequalityVal = parseFloat(document.getElementById("inequalityVal").value);
 	var operator = document.getElementById("operator").value;
     var inequality = document.getElementById("inequality").value;
-	var index = parseInt(document.getElementById("index").value);
+	var index = parseFloat(document.getElementById("index").value);
 	var indexName = document.getElementById("indexName").value;
 
 	// Declare result
